Added Shipment Request, Purchase, Store, and modified Proxy. And added stub functions for all.
